+++ XCDYouTubeKit Demo/iOS Demo/DemoFullScreenViewController.h
@@ -2,7 +2,9 @@
 //  Copyright (c) 2013-2014 Cédric Luthi. All rights reserved.
 //
 
-@interface DemoFullScreenViewController : UIViewController <UITextFieldDelegate>
+#import "VideoPickerController.h"
+
+@interface DemoFullScreenViewController : UIViewController <UITextFieldDelegate, VideoPickerControllerDelegate>
 
 @property (nonatomic, weak) IBOutlet UITextField *videoIdentifierTextField;
 @property (nonatomic, weak) IBOutlet UISwitch *lowQualitySwitch;